+++ Entities/Event.cs
@@ -1,14 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using EventbriteNET.HttpApi;
 
 namespace EventbriteNET.Entities
 {
     public class Event : EntityBase
     {
-        public int Id;
+        public long Id;
         public string Title;
         public string Description;
         public DateTime StartDateTime;
@@ -16,7 +14,7 @@ namespace EventbriteNET.Entities
         public DateTime Created;
         public DateTime Modified;
 
-        public Dictionary<int, Ticket> Tickets = new Dictionary<int, Ticket>();
+        public Dictionary<long, Ticket> Tickets = new Dictionary<long, Ticket>();
 
         private List<Attendee> attendees;
         public List<Attendee> Attendees
