+++ Entities/Ticket.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
 namespace EventbriteNET.Entities
 {
@@ -12,7 +9,7 @@ namespace EventbriteNET.Entities
     }
     public class Ticket : EntityBase
     {
-        public int Id;
+        public long Id;
         public string Name;
         public string Description;
         public TicketType Type;
